Move some functions from LangPolymorphic to new Macros module
