+++ code/drasil-gool/GOOL/Drasil/LanguageRenderer/PythonRenderer.hs
@@ -49,7 +49,7 @@ import GOOL.Drasil.LanguageRenderer.Constructors (mkStmtNoEnd, mkStateVal,
   mkVal, mkStateVar, VSOp, unOpPrec, powerPrec, multPrec, andPrec, orPrec, 
   unExpr, unExpr', typeUnExpr, binExpr, typeBinExpr)
 import qualified GOOL.Drasil.LanguageRenderer.LanguagePolymorphic as G (
-  multiBody, block, multiBlock, int, listInnerType, obj, funcType, runStrategy, 
+  multiBody, block, multiBlock, int, listInnerType, obj, funcType, 
   notOp', negateOp, sqrtOp', absOp', expOp', sinOp', cosOp', tanOp', asinOp', 
   acosOp', atanOp', csc, sec, cot, equalOp, notEqualOp, greaterOp, 
   greaterEqualOp, lessOp, lessEqualOp, plusOp, minusOp, multOp, divideOp, 
@@ -60,15 +60,17 @@ import qualified GOOL.Drasil.LanguageRenderer.LanguagePolymorphic as G (
   extNewObjMixedArgs, lambda, func, get, set, listAdd, listAppend, iterBegin, 
   iterEnd, listAccess, listSet, getFunc, setFunc, listAddFunc, listAppendFunc, 
   iterBeginError, iterEndError, listAccessFunc, listSetFunc, stmt, loopStmt, 
-  emptyStmt, assign, decrement, increment', increment1', decrement1, 
+  emptyStmt, assign, increment, 
   listDecDef', objDecNew, objDecNewNoParams, print, closeFile, discardFileLine, 
-  stringListVals, stringListLists, returnStmt, valStmt, comment, throw, 
+  returnStmt, valStmt, comment, throw, 
   ifCond, ifExists, tryCatch, checkState, construct, param, method, getMethod, 
   setMethod, constructor, function, docFunc, stateVarDef, constVar, buildClass, 
   implementingClass, docClass, commentedClass, intClass, buildModule, 
   modFromData, fileDoc, docMod, fileFromData)
 import GOOL.Drasil.LanguageRenderer.LanguagePolymorphic (addmathImport, 
   bindingError, destructorError, docFuncRepr)
+import qualified GOOL.Drasil.LanguageRenderer.Macros as M (decrement, 
+  decrement1, increment1, runStrategy, stringListVals, stringListLists)
 import GOOL.Drasil.AST (Terminator(..), FileType(..), 
   FileData(..), fileD, FuncData(..), fd, ModData(..), md, updateMod, 
   MethodData(..), mthd, updateMthd, OpData(..), ParamData(..), pd, 
@@ -489,10 +491,10 @@ instance StatementSym PythonCode where
 
 instance AssignStatement PythonCode where
   assign = G.assign Empty
-  (&-=) = G.decrement
-  (&+=) = G.increment'
-  (&++) = G.increment1'
-  (&--) = G.decrement1
+  (&-=) = M.decrement
+  (&+=) = G.increment
+  (&++) = M.increment1
+  (&--) = M.decrement1
 
 instance DeclStatement PythonCode where
   varDec _ = toState $ mkStmtNoEnd empty
@@ -543,8 +545,8 @@ instance StringStatement PythonCode where
   stringSplit d vnew s = assign vnew (objAccess s (func "split" 
     (listType string) [litString [d]]))  
 
-  stringListVals = G.stringListVals
-  stringListLists = G.stringListLists
+  stringListVals = M.stringListVals
+  stringListLists = M.stringListLists
 
 instance FuncAppStatement PythonCode where
   inOutCall = pyInOutCall funcApp
@@ -590,7 +592,7 @@ instance ObserverPattern PythonCode where
           notify = oneLiner $ valStmt $ at obsList (valueOf index) $. f
 
 instance StrategyPattern PythonCode where
-  runStrategy = G.runStrategy
+  runStrategy = M.runStrategy
 
 instance ScopeSym PythonCode where
   type Scope PythonCode = Doc
