+++ code/drasil-gool/drasil-gool.cabal
@@ -17,6 +17,7 @@ library
     , GOOL.Drasil.CodeInfo
     , GOOL.Drasil.LanguageRenderer.Constructors
     , GOOL.Drasil.LanguageRenderer.LanguagePolymorphic
+    , GOOL.Drasil.LanguageRenderer.Macros
     , GOOL.Drasil.LanguageRenderer
     , GOOL.Drasil.LanguageRenderer.JavaRenderer
     , GOOL.Drasil.LanguageRenderer.PythonRenderer
@@ -59,6 +60,7 @@ executable codegenTest
     , GOOL.Drasil.CodeInfo
     , GOOL.Drasil.LanguageRenderer.Constructors
     , GOOL.Drasil.LanguageRenderer.LanguagePolymorphic
+    , GOOL.Drasil.LanguageRenderer.Macros
     , GOOL.Drasil.LanguageRenderer
     , GOOL.Drasil.LanguageRenderer.JavaRenderer
     , GOOL.Drasil.LanguageRenderer.PythonRenderer
